moved middleware out of server.js
