+++ api/server.js
@@ -1,59 +1,14 @@
 const userDb = require('../data/helpers/userDb');
 const postDb = require('../data/helpers/postDb');
+const configureMiddleware = require('../config/middleware');
+const capitalizeName = require('../common/capitalizeName');
+const checkForValidUser = require('../common/checkForValidUser');
 
 const express = require('express');
-const morgan = require('morgan');
-const helmet = require('helmet');
-const cors = require('cors');
 
 const server = express();
 
-// MIDDLEWARE
-server.use(helmet());
-server.use(morgan('dev'));
-server.use(express.json());
-server.use(cors());
-
-// capitalize users name
-const capitalizeName = (req, res, next) => {
-  // expects an object with a 'name' field, ie: {name: 'justin'}
-  if (!req.body.name) {
-    res.status(400).json({errorMessage: 'no name in request'});
-  } else {
-    //req.body.name = req.body.name[0].toUpperCase() + req.body.name.slice(1);
-    let name = req.body.name.split(' ');
-    console.log(name);
-    name = name.map(word => word[0].toUpperCase() + word.slice(1));
-    console.log(name);
-    req.body.name = name.join(' ');
-
-    next();
-  }
-};
-
-// ensure user exists
-const checkForValidUser = (req, res, next) => {
-  const id = req.params.id;
-  userDb
-    .get(id)
-    .then(user => {
-      // user was found, carry on
-      if (user) {
-        next();
-      } else {
-        res
-          .status(400)
-          .json({errorMessage: 'a user with that id does not exist'});
-      }
-    })
-    .catch(err => {
-      // i don't know what this error could be, actually
-      res
-        .status(500)
-        .json({errorMessage: `there was an error retrieving the user: ${err}`});
-    });
-};
-
+configureMiddleware(server);
 // ROUTES
 server.get('/', (req, res) => {
   res.status(200).send('sanity check!');
